training/requirements.txt
# Training Module Requirements
# Install with: pip install -r requirements.txt

# Deep Learning Framework
tensorflow==2.15.0
# OR for Apple Silicon Macs:
# tensorflow-macos==2.15.0
# tensorflow-metal

# Data Processing
numpy==1.26.2
pandas==2.1.3

# Image Processing
Pillow==10.1.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Machine Learning Utilities
scikit-learn==1.3.2

# Progress Bars (optional)
tqdm==4.66.1
